Move dummy alert to alert types factory
